LayoutTests/media/track/text-track-src-change-expected.txt

Tests Track 'src' changing handling



*** Testing Track 'src' changing
EXPECTED (cues.length == '4') OK

*** Changing 'src' attribute...
EXPECTED (cues.length == '100') OK

*** Changing back 'src' attribute...
EXPECTED (cues.length == '4') OK
END OF TEST
